Finish implementing tags
Every node block has tags that range from 0 to the number set in the config. That number can be changed at any time and requires just restarting the game. Tags are changed by crouching and right-clicking. They are visible when looking at the nodes. More work needs to be done to integrate these changes with the test framework.