Fix permissive command parsing
Rewrote functionality of CommandCircuitryBase to utilize Java Reflection a lot more in order to make it easier to create new commands. More work is done in that class and less is done in implementing classes. Also, command argument order and quantity is now enforced properly for optional arguments.

Renamed event handler classes.