Fix /test
Fixed /test failing for no reason when the optional arguments are not specified. Changed delay to specify the time between tests instead of the time to wait between setting and getting the results of a test. Changed the test phase variable to an enum to make adding new phases much easier. Forced Gradle to compile the mod with parameter names included so that the Reflection used in CommandCircuitryBase can properly retrieve the names of command arguments.